🧹 Tidy: [NoteHistoryのFat Component分割と型安全の向上]
1. NoteHistoryの編集ダイアログとフォームロジックを独立したNoteEditDialogとして抽出。
2. 独自のDialogプリミティブをEditDialogBaseに置き換え。
3. NoteHistoryの手動の日付オフセット計算をformatDateTimeLocalに置き換え。
4. HexagonWidgetCard内の冗長な403エラー判定をisApiErrorユーティリティに置き換え。
