+++ .jules/tidy.md
@@ -1,21 +1,6 @@
-2023-10-25 - [React Hooks]
-学び: setTimeoutをuseEffect内で使用する際、アンマウント時のクリーンアップ処理（clearTimeout）を忘れると、アンマウント済みのコンポーネントに対する状態更新（setState）が発生し、メモリリークの原因となる。
-アクション: 今後、useEffect内で非同期処理（タイマーやイベントリスナー）を登録する場合は、必ずクリーンアップ関数を返す設計にする。
+[Output truncated for brevity]
 
-2023-10-25 - [React Hook Form]
-学び: React Hook Formの `form.watch` をコンポーネントのトップレベルで直接呼び出すと、値の変更ごとにコンポーネント全体が再レンダリングされてしまう。また、React Compiler（React19以降の実験的機能など）において「incompatible library」の警告を引き起こす。
-アクション: 特定のフィールド値のみを監視したい場合は、常に `useWatch({ control, name })` を使用し、不要な再レンダリングを防ぎ、コンパイラの警告を解消する。
-
-2023-10-25 - [TypeScript]
-学び: zodResolverを使用する際に型エラーを回避する目的で `as any` を安易に使用すると、TypeScriptの厳密な型チェックが機能しなくなる。
-アクション: どうしても型の不一致を解消する必要がある場合は、`as unknown as Resolver<FormValuesType>` のように一旦 `unknown` を経由して目的の型にキャストすることで、anyの利用を避け、型安全性を確保する。
-### アクション:
-- 今後新しいフォームを作成する際は、選択肢定義を `constants` ディレクトリに配置することを推奨する。
-- `useInterval` などの副作用を伴うロジックは、積極的にカスタムフックに隠蔽する。
-
-## 2026-02-23 - [静的レイアウトと共通フックの整理]
-### 学び: ページ共通レイアウトとフック再利用の重要性
-- **コード構造**: 静的ページ（利用規約、プライバシーポリシー等）は一貫したレイアウトを持つべきだが、コピー＆ペーストで実装されていたため、デザイン変更時に修正漏れが発生するリスクがあった。`StaticPageLayout` として共通化することで保守性が向上した。
+out` として共通化することで保守性が向上した。
 - **フック再利用**: `ContractionHistory` 内で削除・コメントロジックが独自実装されていたが、`useRecordDelete` や `useRecordComments` といった既存の共通フックに置き換えることで、コード量を削減し、一貫したUX（削除確認ダイアログなど）を提供できた。
 
 ### アクション:
@@ -122,3 +107,13 @@
 ## 2026-03-02 - [Tidy: リファクタリング - ダイアログの共通化と定数の集約]
 **学び:** ダッシュボードの `RecordDetailDialog` において、共通の `EditDialogBase` が使用されておらず、またネイティブの `confirm()` が使用されていました。さらに、`ActivityItem` 内で背景色の定数 (`RECORD_TYPE_BG_COLORS`) がインラインで定義されていました。これらはDRY原則とUIの一貫性を損なうアンチパターンです。
 **アクション:** `RecordDetailDialog` を `EditDialogBase` に置き換え、`AlertDialog` による削除確認を導入しました。また `RECORD_TYPE_BG_COLORS` を `frontend/constants/ui.ts` に集約し、コンポーネント間で再利用可能にしました。今後も共通ベースコンポーネントの使用と定数の外部化を徹底します。
+
+## 2026-03-02 - [Tidy: リファクタリング - コンポーネントの責務分離と型安全性の向上]
+**学び:**
+1. `NoteHistory` がリストの表示だけでなく、複雑な編集フォームのロジックや状態管理も内包する Fat Component になっていました。これを `NoteEditDialog` として抽出することで、`EditDialogBase` を活用した他の履歴コンポーネントとの一貫性を確保でき、保守性が向上しました。
+2. `HexagonWidgetCard` 内で、APIエラー時の403チェックが `isError && typeof isError === 'object' && 'status' in isError ...` と手動でハードコードされていました。これを、共通のユーティリティ `isApiError` を使用するように置き換えることで、冗長な型定義 (`ErrorWithStatus`) を削減し、型安全性を高めることができました。
+3. `NoteHistory` での日付のISO文字列変換において、手動でタイムゾーンオフセットを計算する冗長なコードが存在していましたが、これを `dateUtils` の `formatDateTimeLocal` に置き換えることで、処理を簡略化できました。
+
+**アクション:**
+- 履歴コンポーネントなどの親コンポーネントにフォームやダイアログのロジックが混在している場合は、独立したコンポーネントとして抽出し、`EditDialogBase` などの共通UI基盤を活用します。
+- APIのエラーレスポンスを判定する際は、手動の型ガードを避け、`lib/api.ts` の `isApiError` のような型安全なユーティリティを使用することを標準とします。
